+++ CHANGELOG.md
@@ -73,6 +73,10 @@ and this project adheres to
 
 - 🔥(backend) drop deprecated numchild columns from item
 
+### Fixed
+
+- 🐛(backend) prevent moving an item to itself or one of its descendants
+
 ## [v0.18.0] - 2026-05-04
 
 ### Added